feat(HttpInterceptor): do not stop loadingService for silent or back-end calls
When 2 back-end services are called at the same time and one of them is silent, loading service will be stopped as soon as the silent call is over.
For instance, if there'e a back-end get along with a log, then the loading service will be stopped almost instantly as the log is silent and over in no time.
The intended result is that the loading service only stops once the get is over.