fix(master-layout): hide horizontal scrollbar
When only the footer is sticky, the parent of the header must have `overflow: auto`. In that case,
when the main navigation is longer than the available place, an horizontal scrollbar appears.
To remove it, we need an `overflow: hidden` on the navigation itself, but this prevents the
sub navigation to be opened since it would overflowing from the main navigation.
The optimal solution would be to place the sub navigation out of the main navigation but this
is far too complex. Thus the compromise of adding `overflow: hidden` only when the sub navigation
is closed. That way, the unwanted scrollbar only appears when all of the following are true:
* the footer is sticky but the header isn't
* the main navigation is longer than the available
* one of the sub navigation item is opened

OUI-1537