refactor(sds): remove `PreviewComponent` interface
The `PreviewComponent` interface didn't serve its intended purpose. The idea was to have a single
type for all previews, but unfortunately, the `Type<T>` operator used to enforce this don't check
that the interface is present, but merely that all its properties are implemented. As
`PreviewComponent` is empty, `Type<PreviewComponent>` is effectively equal to `Type<void>`.
Removing it simplifies the code and eliminates a source of incoherence while maintaining the exact
same behavior.

OUI-2541