fix(toolchain): use different functions to execute script commands with or without returning results
Commands whose result need to be parsed need to keep their default `stdio`. On the other hand,
commands whose result should simply be outputted in the terminal need to inherit the `stdio`. This
means two different functions are needed. One to get the result from a command, the other one to
execute a command. The second function has an option to output the command and its results to the
terminal.

OUI-3153