refactor(oblique/service-navigation): use interpolation in `calc` to avoid confusing the IDE
`calc($var)` is seen as an error by the IDE because `$var` is not a valid input to the CSS `calc`
function. Using interpolation makes it clear the SCSS `calc` function is to be used instead of the
CSS one. In this case it has no other effect as the SCSS `calc` will be transpiled into the CSS
`calc`.

OUI-3232