refactor(cli/utils): `getVersionedDependency` always returns a versioned dependency
Since this function only handles dependencies listed in `currentVersions`, the only scenario where a
version might be missing is if `currentVersions` explicitly includes an empty string for a
dependency's version. Since this would be an error that needs to be addressed in `currentVersions`,
there's no need to handle this case within `getVersionedDependency`.

OUI-3438