feat(sds/autocomplete): add example with `displayWith`
The coverage had to be reduced because this feature has no associated tests. The current level of
coverage is purely incidental. Achieving the same coverage would require adding a test file for the
new preview, which would be inconsistent with the established practice of not writing tests for
previews.

OUI-4128